Change default number type for `Flat`
Given that the same type is used both for delay and speed, and that
we're computing one from the other, it makes no sense to more accuracy
on one side of the decimal point. This could only lead to problems when
computing the inverse, if that accuracy were actually used to its full
extent.